Backend/requirements.txt
asgiref 
Django 
django-cors-headers 
djangorestframework 
djangorestframework-simplejwt 
PyJWT 
pytz 
sqlparse 
psycopg2-binary 
python-dotenv